WIP registration working, with lots of TODOs and FIXMEs

+++ src/repository/user.go
@@ -1,12 +1,15 @@
 package repository
 
-import "auth/src/model"
+import (
+	"auth/src/model"
+)
 
 type UserRepository interface {
-	CreateUser(usr *model.User) (model.User, error)
-	GetUserByUsername(username string) (model.User, error)
+	CreateUser(usr *model.User) error
+	GetUserByUsername(username string) (*model.User, error)
 	RemoveUserByUsername(username string) error
-	UpdateUser(usr *model.User) (model.User, error)
+	UpdateUser(usr *model.User) (*model.User, error)
+	Exists(username string) bool
 }
 
 type userRepository struct {
@@ -22,36 +25,47 @@ const USER_LOGIN_TYPE = "user_login"
 const USER_LOGOUT_TYPE = "user_logout"
 const USER_DELETE_TYPE = "user_delete"
 
-func (r *userRepository) GetUserByUsername(username string) (usr model.User, err error) {
-	err = r.c.connPool.QueryRow("SELECT id, username, password FROM user WHERE username = $1", username).Scan(&usr.Id, &usr.Username, &usr.Password)
-	if err != nil {
-		return usr, err
+func (r *userRepository) Exists(username string) bool {
+	if _, err := r.GetUserByUsername(username); err != nil {
+		return false
 	}
-	return usr, nil
+	return true
+}
+
+func (r *userRepository) GetUserByUsername(username string) (usr *model.User, err error) {
+	tmp := model.User{}
+	err = r.c.connPool.QueryRow("SELECT id, username, password FROM auth.user WHERE username = $1", username).Scan(&tmp.Id, &tmp.Username, &tmp.Password)
+	if err != nil {
+		return nil, err
+	}
+	return &tmp, nil
 }
 
 func (r *userRepository) RemoveUserByUsername(usrName string) error {
 	// TODO create event
-	_, err := r.c.connPool.Exec("DELETE FROM user WHERE username = $1", usrName)
+	_, err := r.c.connPool.Exec("DELETE FROM user WHERE auth.username = $1", usrName)
 	return err
 }
 
-func (r *userRepository) UpdateUser(usr *model.User) (updatedUsr model.User, err error) {
+func (r *userRepository) UpdateUser(usr *model.User) (updatedUsr *model.User, err error) {
 	// TODO create event
-	_, err = r.c.connPool.Exec("UPDATE user SET username = $1, password = $2 WHERE id = $3", usr.Username, usr.Password, usr.Id)
+	_, err = r.c.connPool.Exec("UPDATE auth.user SET username = $1, password = $2 WHERE id = $3", usr.Username, usr.Password, usr.Id)
 	return updatedUsr, err // FIXME
 }
 
 // FIXME: wrap in single transaction
-func (r *userRepository) CreateUser(usr *model.User) (newUsr model.User, err error) {
-	row := r.c.connPool.QueryRow("INSERT INTO events (uuid, type, body) VALUES ($1, $2, $3)",
-		usr.Id, USER_CREATE_TYPE, usr) // FIXME remove password from event body
-	if err = row.Err(); err != nil {
-		return *usr, err
+func (r *userRepository) CreateUser(usr *model.User) error {
+	connPool := r.c.connPool
+	// row := r.c.connPool.QueryRow("INSERT INTO auth.events (uuid, type, body) VALUES ($1, $2, $3)",
+	// 	usr.Id, USER_CREATE_TYPE, usr) // FIXME remove password from event body
+	// if err = row.Err(); err != nil {
+	// 	return usr, err
+	// }
+	stmt, err := connPool.Prepare("INSERT INTO auth.user (id, username, password) VALUES ($1, $2, $3)")
+	if err != nil {
+		return err
 	}
-	row = r.c.connPool.QueryRow("INSERT INTO user (id, username, password) VALUES ($1, $2, $3)", usr.Id, usr.Username, usr.Password)
-	if err = row.Err(); err != nil {
-		return *usr, err
-	}
-	return newUsr, nil // FIXME return user from db
+	defer stmt.Close() // https://go.dev/doc/database/prepared-statements
+	_, err = stmt.Exec(usr.Id, usr.Username, usr.Password)
+	return err
 }
